requirements.txt
deepagents>=0.1.0
langchain>=0.3.0
langgraph>=0.3.0
langchain-anthropic>=0.3.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
